deployment/backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
numpy
pandas
scikit-learn
xgboost
joblib
pillow
shap
lime
matplotlib
sqlalchemy
psycopg2-binary
